Joined all memoization policy tests
