+++ tests/core/test_policies.py
@@ -573,6 +573,73 @@ class TestMemoizationPolicy(PolicyTestCollection):
 
         assert isinstance(featurizer.state_featurizer, state_featurizer)
 
+    @pytest.mark.parametrize("max_history", [1, 2, 3, 4, None])
+    def test_prediction(
+        self,
+        max_history: Optional[int],
+        model_storage: ModelStorage,
+        resource: Resource,
+        execution_context: ExecutionContext,
+    ):
+        policy = self.create_policy(
+            featurizer=MaxHistoryTrackerFeaturizer(max_history=max_history),
+            model_storage=model_storage,
+            resource=resource,
+            execution_context=execution_context,
+        )
+
+        GREET_INTENT_NAME = "greet"
+        UTTER_GREET_ACTION = "utter_greet"
+        UTTER_BYE_ACTION = "utter_goodbye"
+        domain = Domain.from_yaml(
+            f"""
+            intents:
+            - {GREET_INTENT_NAME}
+            actions:
+            - {UTTER_GREET_ACTION}
+            - {UTTER_BYE_ACTION}
+            slots:
+                slot_1:
+                    type: bool
+                slot_2:
+                    type: bool
+                slot_3:
+                    type: bool
+                slot_4:
+                    type: bool
+            """
+        )
+        events = [
+            UserUttered(intent={"name": GREET_INTENT_NAME}),
+            ActionExecuted(UTTER_GREET_ACTION),
+            SlotSet("slot_1", True),
+            ActionExecuted(UTTER_GREET_ACTION),
+            SlotSet("slot_2", True),
+            SlotSet("slot_3", True),
+            ActionExecuted(UTTER_GREET_ACTION),
+            ActionExecuted(UTTER_GREET_ACTION),
+            UserUttered(intent={"name": GREET_INTENT_NAME}),
+            ActionExecuted(UTTER_GREET_ACTION),
+            SlotSet("slot_4", True),
+            ActionExecuted(UTTER_BYE_ACTION),
+        ]
+        training_story = TrackerWithCachedStates.from_events(
+            "training story", evts=events, domain=domain, slots=domain.slots,
+        )
+        test_story = TrackerWithCachedStates.from_events(
+            "training story", events[:-1], domain=domain, slots=domain.slots,
+        )
+        policy.train([training_story], domain)
+        prediction = policy.predict_action_probabilities(
+            test_story, domain, RegexInterpreter()
+        )
+        assert (
+            domain.action_names_or_texts[
+                prediction.probabilities.index(max(prediction.probabilities))
+            ]
+            == UTTER_BYE_ACTION
+        )
+
 
 class TestAugmentedMemoizationPolicy(TestMemoizationPolicy):
     """Test suite for AugmentedMemoizationPolicy."""
@@ -580,6 +647,84 @@ class TestAugmentedMemoizationPolicy(TestMemoizationPolicy):
     @staticmethod
     def _policy_class_to_test() -> Type[PolicyGraphComponent]:
         return AugmentedMemoizationPolicy
+
+    @pytest.mark.parametrize("max_history", [1, 2, 3, 4, None])
+    def test_augmented_prediction(
+        self,
+        max_history: Optional[int],
+        model_storage: ModelStorage,
+        resource: Resource,
+        execution_context: ExecutionContext,
+    ):
+        policy = self.create_policy(
+            featurizer=MaxHistoryTrackerFeaturizer(max_history=max_history),
+            model_storage=model_storage,
+            resource=resource,
+            execution_context=execution_context,
+        )
+
+        GREET_INTENT_NAME = "greet"
+        UTTER_GREET_ACTION = "utter_greet"
+        UTTER_BYE_ACTION = "utter_goodbye"
+        domain = Domain.from_yaml(
+            f"""
+                intents:
+                - {GREET_INTENT_NAME}
+                actions:
+                - {UTTER_GREET_ACTION}
+                - {UTTER_BYE_ACTION}
+                slots:
+                    slot_1:
+                        type: bool
+                        initial_value: true
+                    slot_2:
+                        type: bool
+                    slot_3:
+                        type: bool
+                """
+        )
+        training_story = TrackerWithCachedStates.from_events(
+            "training story",
+            [
+                ActionExecuted(UTTER_GREET_ACTION),
+                UserUttered(intent={"name": GREET_INTENT_NAME}),
+                ActionExecuted(UTTER_GREET_ACTION),
+                SlotSet("slot_3", True),
+                ActionExecuted(UTTER_BYE_ACTION),
+            ],
+            domain=domain,
+            slots=domain.slots,
+        )
+        test_story = TrackerWithCachedStates.from_events(
+            "test story",
+            [
+                UserUttered(intent={"name": GREET_INTENT_NAME}),
+                ActionExecuted(UTTER_GREET_ACTION),
+                SlotSet("slot_1", False),
+                ActionExecuted(UTTER_GREET_ACTION),
+                ActionExecuted(UTTER_GREET_ACTION),
+                UserUttered(intent={"name": GREET_INTENT_NAME}),
+                ActionExecuted(UTTER_GREET_ACTION),
+                SlotSet("slot_2", True),
+                ActionExecuted(UTTER_GREET_ACTION),
+                UserUttered(intent={"name": GREET_INTENT_NAME}),
+                ActionExecuted(UTTER_GREET_ACTION),
+                SlotSet("slot_3", True),
+                # ActionExecuted(UTTER_BYE_ACTION),
+            ],
+            domain=domain,
+            slots=domain.slots,
+        )
+        policy.train([training_story], domain)
+        prediction = policy.predict_action_probabilities(
+            test_story, domain, RegexInterpreter()
+        )
+        assert (
+            domain.action_names_or_texts[
+                prediction.probabilities.index(max(prediction.probabilities))
+            ]
+            == UTTER_BYE_ACTION
+        )
 
 
 @pytest.mark.parametrize(
